Split out the parser logic into its own file, now we only export a single function. A few renames too
