+++ src/parser.spec.ts
@@ -1,32 +1,31 @@
 import { readFileSync } from 'fs';
-import { OpenLyricsSong } from './model';
-import { OpenLyrics } from '.';
+import { IOpenLyricsSong } from './parser.model';
+import { OpenLyricsParser } from '.';
 
-describe('OpenLyrics', (): void => {
-  let olParser: OpenLyrics;
+//NOTE:
+// This test file will test the `OpenLyricsParser` method on the `index.ts` file
+// It does not directly test anything in `parser.ts` even though the name of this file
+// might seem to indicate otherwise
 
-  beforeEach(() => {
-    olParser = new OpenLyrics();
-  });
-
+describe('OpenLyricsParser', (): void => {
   it('should exist', () => {
-    expect(olParser).toBeDefined();
+    expect(OpenLyricsParser).toBeDefined();
   });
 
   describe('Example Files', () => {
     it('should return a song for file: simple.xml"', () => {
       const testFile = readFileSync('./sample-files/examples/simple.xml').toString();
-      const parsedSong = olParser.parse(testFile);
+      const parsedSong = OpenLyricsParser(testFile);
 
-      expect(parsedSong.meta).toEqual<OpenLyricsSong.IMeta>({
+      expect(parsedSong.meta).toEqual<IOpenLyricsSong.IMeta>({
         createdIn: 'OpenLP 1.9.0',
         chordNotation: '',
         modifiedDate: new Date('2012-04-10T12:00:00.000Z'),
         modifiedIn: 'MyApp 0.0.1',
         version: '0.8',
       });
-      expect(parsedSong.format).toEqual<OpenLyricsSong.IFormat>({ application: '', tags: [] });
-      expect(parsedSong.properties).toEqual<OpenLyricsSong.IProperties>({
+      expect(parsedSong.format).toEqual<IOpenLyricsSong.IFormat>({ application: '', tags: [] });
+      expect(parsedSong.properties).toEqual<IOpenLyricsSong.IProperties>({
         authors: [],
         ccliNo: '',
         comments: [],
@@ -45,7 +44,7 @@ describe('OpenLyrics', (): void => {
         verseOrder: '',
         version: '',
       });
-      expect(parsedSong.verses).toEqual<OpenLyricsSong.IVerse[]>([
+      expect(parsedSong.verses).toEqual<IOpenLyricsSong.IVerse[]>([
         {
           name: 'v1',
           transliteration: '',
@@ -63,22 +62,22 @@ describe('OpenLyrics', (): void => {
           ],
         },
       ]);
-      expect(parsedSong.instruments).toEqual<OpenLyricsSong.IInstrument[]>([]);
+      expect(parsedSong.instruments).toEqual<IOpenLyricsSong.IInstrument[]>([]);
     });
 
     it('should return a song for file: complex.xml"', () => {
       const testFile = readFileSync('./sample-files/examples/complex.xml').toString();
-      const parsedSong = olParser.parse(testFile);
+      const parsedSong = OpenLyricsParser(testFile);
 
-      expect(parsedSong.meta).toEqual<OpenLyricsSong.IMeta>({
+      expect(parsedSong.meta).toEqual<IOpenLyricsSong.IMeta>({
         createdIn: 'OpenLP 1.9.0',
         chordNotation: '',
         modifiedDate: new Date('2012-04-10T12:00:00.000Z'),
         modifiedIn: 'ChangingSong 0.0.1',
         version: '0.8',
       });
-      expect(parsedSong.format).toEqual<OpenLyricsSong.IFormat>({ application: '', tags: [] });
-      expect(parsedSong.properties).toEqual<OpenLyricsSong.IProperties>({
+      expect(parsedSong.format).toEqual<IOpenLyricsSong.IFormat>({ application: '', tags: [] });
+      expect(parsedSong.properties).toEqual<IOpenLyricsSong.IProperties>({
         titles: [
           { lang: 'en-US', original: true, value: 'Amazing Grace' },
           { lang: 'en', original: null, value: 'Amazing Grace' },
@@ -119,7 +118,7 @@ describe('OpenLyrics', (): void => {
         verseOrder: 'v1 v2  v3 c v4 c1 c2 b b1 b2',
         version: '0.99',
       });
-      expect(parsedSong.verses).toEqual<OpenLyricsSong.IVerse[]>([
+      expect(parsedSong.verses).toEqual<IOpenLyricsSong.IVerse[]>([
         {
           name: 'v1',
           transliteration: '',
@@ -226,21 +225,21 @@ describe('OpenLyrics', (): void => {
           ],
         },
       ]);
-      expect(parsedSong.instruments).toEqual<OpenLyricsSong.IInstrument[]>([]);
+      expect(parsedSong.instruments).toEqual<IOpenLyricsSong.IInstrument[]>([]);
     });
 
     it('should return a song with format tags for file: format.xml"', () => {
       const testFile = readFileSync('./sample-files/examples/format.xml').toString();
-      const parsedSong = olParser.parse(testFile);
+      const parsedSong = OpenLyricsParser(testFile);
 
-      expect(parsedSong.meta).toEqual<OpenLyricsSong.IMeta>({
+      expect(parsedSong.meta).toEqual<IOpenLyricsSong.IMeta>({
         createdIn: 'OpenLP 1.9.0',
         chordNotation: '',
         modifiedDate: new Date('2012-04-10T12:00:00.000Z'),
         modifiedIn: 'OpenLP 1.9.7',
         version: '0.8',
       });
-      expect(parsedSong.format).toEqual<OpenLyricsSong.IFormat>({
+      expect(parsedSong.format).toEqual<IOpenLyricsSong.IFormat>({
         application: 'OpenLP',
         tags: [
           {
@@ -255,7 +254,7 @@ describe('OpenLyrics', (): void => {
           },
         ],
       });
-      expect(parsedSong.properties).toEqual<OpenLyricsSong.IProperties>({
+      expect(parsedSong.properties).toEqual<IOpenLyricsSong.IProperties>({
         authors: [],
         ccliNo: '',
         comments: [],
@@ -274,7 +273,7 @@ describe('OpenLyrics', (): void => {
         verseOrder: '',
         version: '',
       });
-      expect(parsedSong.verses).toEqual<OpenLyricsSong.IVerse[]>([
+      expect(parsedSong.verses).toEqual<IOpenLyricsSong.IVerse[]>([
         {
           name: 'v1',
           transliteration: '',
@@ -302,21 +301,21 @@ describe('OpenLyrics', (): void => {
           ],
         },
       ]);
-      expect(parsedSong.instruments).toEqual<OpenLyricsSong.IInstrument[]>([]);
+      expect(parsedSong.instruments).toEqual<IOpenLyricsSong.IInstrument[]>([]);
     });
 
     it('should return a song with format tags for file: format2.xml"', () => {
       const testFile = readFileSync('./sample-files/examples/format2.xml').toString();
-      const parsedSong = olParser.parse(testFile);
+      const parsedSong = OpenLyricsParser(testFile);
 
-      expect(parsedSong.meta).toEqual<OpenLyricsSong.IMeta>({
+      expect(parsedSong.meta).toEqual<IOpenLyricsSong.IMeta>({
         createdIn: 'OpenLP 1.9.7',
         chordNotation: '',
         modifiedDate: new Date('2012-04-10T12:00:00.000Z'),
         modifiedIn: 'OpenLP 1.9.7',
         version: '0.8',
       });
-      expect(parsedSong.format).toEqual<OpenLyricsSong.IFormat>({
+      expect(parsedSong.format).toEqual<IOpenLyricsSong.IFormat>({
         application: 'OpenLP',
         tags: [
           {
@@ -366,7 +365,7 @@ describe('OpenLyrics', (): void => {
           },
         ],
       });
-      expect(parsedSong.properties).toEqual<OpenLyricsSong.IProperties>({
+      expect(parsedSong.properties).toEqual<IOpenLyricsSong.IProperties>({
         authors: [{ lang: '', type: '', value: 'M. Jan Hus' }],
         ccliNo: '',
         comments: [],
@@ -385,7 +384,7 @@ describe('OpenLyrics', (): void => {
         verseOrder: '',
         version: '',
       });
-      expect(parsedSong.verses).toEqual<OpenLyricsSong.IVerse[]>([
+      expect(parsedSong.verses).toEqual<IOpenLyricsSong.IVerse[]>([
         {
           name: 'v1',
           transliteration: '',
@@ -506,25 +505,25 @@ describe('OpenLyrics', (): void => {
           ],
         },
       ]);
-      expect(parsedSong.instruments).toEqual<OpenLyricsSong.IInstrument[]>([]);
+      expect(parsedSong.instruments).toEqual<IOpenLyricsSong.IInstrument[]>([]);
     });
 
     it('should return a song with instrument tags in the lyrics for file: laboratory.xml"', () => {
       const testFile = readFileSync('./sample-files/examples/laboratory.xml').toString();
-      const parsedSong = olParser.parse(testFile);
+      const parsedSong = OpenLyricsParser(testFile);
 
-      expect(parsedSong.meta).toEqual<OpenLyricsSong.IMeta>({
+      expect(parsedSong.meta).toEqual<IOpenLyricsSong.IMeta>({
         createdIn: '',
         chordNotation: '',
         modifiedDate: null,
         modifiedIn: '',
         version: '0.9',
       });
-      expect(parsedSong.format).toEqual<OpenLyricsSong.IFormat>({
+      expect(parsedSong.format).toEqual<IOpenLyricsSong.IFormat>({
         application: '',
         tags: [],
       });
-      expect(parsedSong.properties).toEqual<OpenLyricsSong.IProperties>({
+      expect(parsedSong.properties).toEqual<IOpenLyricsSong.IProperties>({
         authors: [{ lang: '', type: '', value: 'Gellért Gyuris' }],
         ccliNo: '',
         comments: [],
@@ -543,7 +542,7 @@ describe('OpenLyrics', (): void => {
         verseOrder: 'i v1 v2',
         version: '',
       });
-      expect(parsedSong.verses).toEqual<OpenLyricsSong.IVerse[]>([
+      expect(parsedSong.verses).toEqual<IOpenLyricsSong.IVerse[]>([
         {
           lang: '',
           lines: [
@@ -641,7 +640,7 @@ describe('OpenLyrics', (): void => {
           transliteration: '',
         },
       ]);
-      expect(parsedSong.instruments).toEqual<OpenLyricsSong.IInstrument[]>([
+      expect(parsedSong.instruments).toEqual<IOpenLyricsSong.IInstrument[]>([
         {
           lines: [
             {
@@ -680,20 +679,20 @@ describe('OpenLyrics', (): void => {
 
     it('should return a song for file: test0.9.xml"', () => {
       const testFile = readFileSync('./sample-files/examples/test0.9.xml').toString();
-      const parsedSong = olParser.parse(testFile);
+      const parsedSong = OpenLyricsParser(testFile);
 
-      expect(parsedSong.meta).toEqual<OpenLyricsSong.IMeta>({
+      expect(parsedSong.meta).toEqual<IOpenLyricsSong.IMeta>({
         createdIn: '',
         chordNotation: 'hungarian',
         modifiedDate: null,
         modifiedIn: '',
         version: '0.9',
       });
-      expect(parsedSong.format).toEqual<OpenLyricsSong.IFormat>({
+      expect(parsedSong.format).toEqual<IOpenLyricsSong.IFormat>({
         application: '',
         tags: [],
       });
-      expect(parsedSong.properties).toEqual<OpenLyricsSong.IProperties>({
+      expect(parsedSong.properties).toEqual<IOpenLyricsSong.IProperties>({
         authors: [
           { lang: '', type: '', value: 'Csiszér László' },
           { lang: '', type: 'words', value: 'Flach Ferenc' },
@@ -718,7 +717,7 @@ describe('OpenLyrics', (): void => {
         verseOrder: 'i v1',
         version: '',
       });
-      expect(parsedSong.verses).toEqual<OpenLyricsSong.IVerse[]>([
+      expect(parsedSong.verses).toEqual<IOpenLyricsSong.IVerse[]>([
         {
           lang: '',
           lines: [
@@ -731,7 +730,7 @@ describe('OpenLyrics', (): void => {
           transliteration: '',
         },
       ]);
-      expect(parsedSong.instruments).toEqual<OpenLyricsSong.IInstrument[]>([
+      expect(parsedSong.instruments).toEqual<IOpenLyricsSong.IInstrument[]>([
         {
           lines: [
             {
@@ -869,20 +868,20 @@ describe('OpenLyrics', (): void => {
 
     it('should return a song for file: version0.9.xml"', () => {
       const testFile = readFileSync('./sample-files/examples/version0.9.xml').toString();
-      const parsedSong = olParser.parse(testFile);
+      const parsedSong = OpenLyricsParser(testFile);
 
-      expect(parsedSong.meta).toEqual<OpenLyricsSong.IMeta>({
+      expect(parsedSong.meta).toEqual<IOpenLyricsSong.IMeta>({
         createdIn: '',
         chordNotation: 'hungarian',
         modifiedDate: null,
         modifiedIn: '',
         version: '0.9',
       });
-      expect(parsedSong.format).toEqual<OpenLyricsSong.IFormat>({
+      expect(parsedSong.format).toEqual<IOpenLyricsSong.IFormat>({
         application: '',
         tags: [],
       });
-      expect(parsedSong.properties).toEqual<OpenLyricsSong.IProperties>({
+      expect(parsedSong.properties).toEqual<IOpenLyricsSong.IProperties>({
         authors: [{ lang: '', type: '', value: 'ismeretlen' }],
         ccliNo: '',
         comments: [],
@@ -901,7 +900,7 @@ describe('OpenLyrics', (): void => {
         verseOrder: 'i v1',
         version: '',
       });
-      expect(parsedSong.verses).toEqual<OpenLyricsSong.IVerse[]>([
+      expect(parsedSong.verses).toEqual<IOpenLyricsSong.IVerse[]>([
         {
           lang: '',
           lines: [
@@ -929,7 +928,7 @@ describe('OpenLyrics', (): void => {
           transliteration: '',
         },
       ]);
-      expect(parsedSong.instruments).toEqual<OpenLyricsSong.IInstrument[]>([
+      expect(parsedSong.instruments).toEqual<IOpenLyricsSong.IInstrument[]>([
         {
           lines: [
             {
@@ -968,20 +967,20 @@ describe('OpenLyrics', (): void => {
       const testFile = readFileSync(
         './sample-files/songs/A Mighty Fortress is Our God.xml'
       ).toString();
-      const parsedSong = olParser.parse(testFile);
+      const parsedSong = OpenLyricsParser(testFile);
 
-      expect(parsedSong.meta).toEqual<OpenLyricsSong.IMeta>({
+      expect(parsedSong.meta).toEqual<IOpenLyricsSong.IMeta>({
         createdIn: 'opensong2openlyrics.py 0.3',
         chordNotation: '',
         modifiedDate: new Date('2012-04-10T21:31:48.233581'),
         modifiedIn: 'convert-schema.py',
         version: '0.9',
       });
-      expect(parsedSong.format).toEqual<OpenLyricsSong.IFormat>({
+      expect(parsedSong.format).toEqual<IOpenLyricsSong.IFormat>({
         application: '',
         tags: [],
       });
-      expect(parsedSong.properties).toEqual<OpenLyricsSong.IProperties>({
+      expect(parsedSong.properties).toEqual<IOpenLyricsSong.IProperties>({
         authors: [{ lang: '', type: '', value: 'Martin Luther' }],
         ccliNo: '',
         comments: [],
@@ -1003,7 +1002,7 @@ describe('OpenLyrics', (): void => {
         verseOrder: 'v1 v2 v3 v4',
         version: '',
       });
-      expect(parsedSong.verses).toEqual<OpenLyricsSong.IVerse[]>([
+      expect(parsedSong.verses).toEqual<IOpenLyricsSong.IVerse[]>([
         {
           lang: '',
           lines: [
@@ -1347,25 +1346,25 @@ describe('OpenLyrics', (): void => {
           transliteration: '',
         },
       ]);
-      expect(parsedSong.instruments).toEqual<OpenLyricsSong.IInstrument[]>([]);
+      expect(parsedSong.instruments).toEqual<IOpenLyricsSong.IInstrument[]>([]);
     });
 
     it('should return a song for the HEBREW file: Hava Nagila.xml"', () => {
       const testFile = readFileSync('./sample-files/songs/Hava Nagila.xml').toString();
-      const parsedSong = olParser.parse(testFile);
+      const parsedSong = OpenLyricsParser(testFile);
 
-      expect(parsedSong.meta).toEqual<OpenLyricsSong.IMeta>({
+      expect(parsedSong.meta).toEqual<IOpenLyricsSong.IMeta>({
         createdIn: 'Trac 0.11.2',
         chordNotation: '',
         modifiedDate: new Date('2012-04-10T21:31:49.006882'),
         modifiedIn: 'convert-schema.py',
         version: '0.9',
       });
-      expect(parsedSong.format).toEqual<OpenLyricsSong.IFormat>({
+      expect(parsedSong.format).toEqual<IOpenLyricsSong.IFormat>({
         application: '',
         tags: [],
       });
-      expect(parsedSong.properties).toEqual<OpenLyricsSong.IProperties>({
+      expect(parsedSong.properties).toEqual<IOpenLyricsSong.IProperties>({
         authors: [],
         ccliNo: '',
         comments: [],
@@ -1398,7 +1397,7 @@ describe('OpenLyrics', (): void => {
       //NOTE: The below Hebrew characters have an invisible control
       //  character to switch the text direction to RTL!!!
       //  Had to use string template quotes to work around this in testing
-      expect(parsedSong.verses).toEqual<OpenLyricsSong.IVerse[]>([
+      expect(parsedSong.verses).toEqual<IOpenLyricsSong.IVerse[]>([
         {
           lang: 'he',
           lines: [
@@ -1563,13 +1562,13 @@ describe('OpenLyrics', (): void => {
           transliteration: '',
         },
       ]);
-      expect(parsedSong.instruments).toEqual<OpenLyricsSong.IInstrument[]>([]);
+      expect(parsedSong.instruments).toEqual<IOpenLyricsSong.IInstrument[]>([]);
     });
   });
 
   // it('should return a song for file: XXX.xml"', () => {
   //   const testFile = readFileSync('./sample-files/songs/XXX.xml').toString();
-  //   const parsedSong = olParser.parse(testFile);
+  //   const parsedSong = OpenLyricsParser(testFile);
 
   //   expect(parsedSong.meta).toEqual<ol.IMeta>({
   //     createdIn: '',
